backend/requirements.txt
asgiref==3.4.1
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
mysqlclient==2.0.3
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
